HDR: Replace expand gamut toggle with multi-mode gamut selection
Change the HDR expand gamut setting from a boolean on/off to a
multi-option uint supporting four colour gamut modes:

- ACCURATE (0): Standard Rec.709 to Rec.2020 conversion
- EXPANDED (1): Microsoft expanded gamut for enhanced saturation
- WIDE (2): Rec.709 to Rec.2020 then to DCI-P3
- SUPER (3): Direct passthrough (no colour space conversion)

Refactor HDR shader pipeline across D3D11, D3D12 and Vulkan to
separate colour space conversion (To2020) from inverse tonemapping
(HDR) and PQ encoding (HDR10), enabling each gamut mode to be
applied at the correct stage. Add DCI-P3 colour matrix (k2020toP3).

Rework scanline processing to operate in gamma 2.4 signal domain
with a dedicated LinearToSignal function, and tune CRT scanline
parameters (max 0.70->0.90, attack 0.30->0.60). Update gamma from
2.2 to 2.22 for more accurate sRGB decoding.

Update menu, configuration, msg_hash, video driver interface,
thread wrapper, and slang reflection to use unsigned type throughout.
Rebuild SPIR-V shader binaries.

+++ gfx/drivers_shader/shader_vulkan.h
@@ -166,7 +166,7 @@ void vulkan_filter_chain_set_max_nits(vulkan_filter_chain_t *chain,
       float max_nits) ;
 
 void vulkan_filter_chain_set_expand_gamut(vulkan_filter_chain_t *chain,
-      float expand_gamut);
+      unsigned expand_gamut);
 
 void vulkan_filter_chain_set_scanlines(vulkan_filter_chain_t *chain,
       float scanlines);
